[REFACTOR] refactor domain.tag
