+++ src/main/java/com/sullung2yo/seatcatcher/domain/tag/dto/response/TagResponse.java
@@ -1,6 +1,6 @@
-package com.sullung2yo.seatcatcher.user.dto.response;
+package com.sullung2yo.seatcatcher.domain.tag.dto.response;
 
-import com.sullung2yo.seatcatcher.user.domain.UserTagType;
+import com.sullung2yo.seatcatcher.domain.tag.enums.UserTagType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
