+++ src/main/java/com/sullung2yo/seatcatcher/domain/tag/entity/Tag.java
@@ -1,7 +1,8 @@
-package com.sullung2yo.seatcatcher.user.domain;
+package com.sullung2yo.seatcatcher.domain.tag.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.sullung2yo.seatcatcher.common.domain.BaseEntity;
+import com.sullung2yo.seatcatcher.domain.tag.enums.UserTagType;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.ColumnDefault;
