+++ src/main/java/com/sullung2yo/seatcatcher/train/dto/response/SeatOccupant.java
@@ -1,8 +1,7 @@
 package com.sullung2yo.seatcatcher.train.dto.response;
 
 import com.sullung2yo.seatcatcher.user.domain.ProfileImageNum;
-import com.sullung2yo.seatcatcher.user.domain.UserTagType;
-import io.swagger.v3.oas.annotations.media.Schema;
+import com.sullung2yo.seatcatcher.domain.tag.enums.UserTagType;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
