+++ src/main/java/com/sullung2yo/seatcatcher/user/controller/UserController.java
@@ -3,7 +3,7 @@ package com.sullung2yo.seatcatcher.user.controller;
 import com.sullung2yo.seatcatcher.common.exception.ErrorCode;
 import com.sullung2yo.seatcatcher.common.exception.TokenException;
 import com.sullung2yo.seatcatcher.user.domain.User;
-import com.sullung2yo.seatcatcher.user.domain.UserTagType;
+import com.sullung2yo.seatcatcher.domain.tag.enums.UserTagType;
 import com.sullung2yo.seatcatcher.user.dto.request.UserInformationUpdateRequest;
 import com.sullung2yo.seatcatcher.user.dto.response.UserInformationResponse;
 import com.sullung2yo.seatcatcher.user.service.UserService;
