+++ src/main/java/com/sullung2yo/seatcatcher/user/domain/User.java
@@ -3,6 +3,7 @@ package com.sullung2yo.seatcatcher.user.domain;
 import com.sullung2yo.seatcatcher.common.domain.BaseEntity;
 import com.sullung2yo.seatcatcher.domain.auth.entity.RefreshToken;
 import com.sullung2yo.seatcatcher.domain.auth.enums.Provider;
+import com.sullung2yo.seatcatcher.domain.tag.entity.UserTag;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Min;
 import lombok.*;
