+++ src/main/java/com/sullung2yo/seatcatcher/user/dto/request/UserInformationUpdateRequest.java
@@ -1,7 +1,7 @@
 package com.sullung2yo.seatcatcher.user.dto.request;
 
 import com.sullung2yo.seatcatcher.user.domain.ProfileImageNum;
-import com.sullung2yo.seatcatcher.user.domain.UserTagType;
+import com.sullung2yo.seatcatcher.domain.tag.enums.UserTagType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
