+++ src/main/java/com/sullung2yo/seatcatcher/user/service/UserServiceImpl.java
@@ -8,11 +8,14 @@ import com.sullung2yo.seatcatcher.common.domain.TokenType;
 import com.sullung2yo.seatcatcher.common.jwt.provider.JwtTokenProviderImpl;
 import com.sullung2yo.seatcatcher.domain.auth.enums.Provider;
 import com.sullung2yo.seatcatcher.domain.auth.service.AuthServiceImpl;
+import com.sullung2yo.seatcatcher.domain.tag.entity.Tag;
+import com.sullung2yo.seatcatcher.domain.tag.entity.UserTag;
+import com.sullung2yo.seatcatcher.domain.tag.enums.UserTagType;
 import com.sullung2yo.seatcatcher.user.domain.*;
 import com.sullung2yo.seatcatcher.user.dto.request.UserInformationUpdateRequest;
-import com.sullung2yo.seatcatcher.user.repository.TagRepository;
+import com.sullung2yo.seatcatcher.domain.tag.repository.TagRepository;
 import com.sullung2yo.seatcatcher.user.repository.UserRepository;
-import com.sullung2yo.seatcatcher.user.repository.UserTagRepository;
+import com.sullung2yo.seatcatcher.domain.tag.repository.UserTagRepository;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
