+++ src/test/java/com/sullung2yo/seatcatcher/user/service/UserServiceImplTest.java
@@ -3,11 +3,14 @@ package com.sullung2yo.seatcatcher.user.service;
 import com.sullung2yo.seatcatcher.common.domain.TokenType;
 import com.sullung2yo.seatcatcher.common.jwt.provider.JwtTokenProviderImpl;
 import com.sullung2yo.seatcatcher.domain.auth.enums.Provider;
+import com.sullung2yo.seatcatcher.domain.tag.entity.Tag;
+import com.sullung2yo.seatcatcher.domain.tag.entity.UserTag;
+import com.sullung2yo.seatcatcher.domain.tag.enums.UserTagType;
 import com.sullung2yo.seatcatcher.user.domain.*;
 import com.sullung2yo.seatcatcher.user.dto.request.UserInformationUpdateRequest;
-import com.sullung2yo.seatcatcher.user.repository.TagRepository;
+import com.sullung2yo.seatcatcher.domain.tag.repository.TagRepository;
 import com.sullung2yo.seatcatcher.user.repository.UserRepository;
-import com.sullung2yo.seatcatcher.user.repository.UserTagRepository;
+import com.sullung2yo.seatcatcher.domain.tag.repository.UserTagRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
